# src/main/resources/openapi.yaml
openapi: 3.0.4
info:
  title: Smart Windows
  description: |-
    Smart Windows - это приложение на базе IoT, предназначенное для того, чтобы пользователи могли удаленно управлять своими умными окнами с помощью мобильного или веб-приложения.

    Полезные ссылки:
    - [Репозиторий Smart Windows на GitHub](https://github.com/dankotyt/smart-windows)

  contact:
    email: danilkotlarov1@gmail.com
  version: 0.0.1
servers:
  - url: http://localhost:8080/api/v1/
    description: dev server
...
paths:
  /analytics/presets:
    post:
      tags:
        - analytics
      summary: Создать новый пресет
      description: Создает новую запись аналитики для пресета
      operationId: createPreset
      requestBody:
        description: Данные пресета в формате JSON
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresetAnalyticsDTO'
      responses:
        '200':
          description: Пресет успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresetAnalyticsDTO'
        '400':
          description: Некорректные данные пресета
  /analytics/presets/downloads:
    post:
      tags:
        - analytics
      summary: Зарегистрировать загрузку пресета
      description: Регистрирует факт загрузки пресета с заданным названием
      operationId: recordPresetDownload
      requestBody:
        description: Название пресета в формате JSON
        required: true
        content:
          application/json:
            schema:
              type: string
              example: "Morning_Mode"
      responses:
        '200':
          description: Загрузка зарегистрирована
          content:
            application/json:
              schema:
                type: string
                format: date-time
        '400':
          description: Некорректные данные
  /analytics/users:
    post:
      tags:
        - analytics
      summary: Создать нового пользователя
      description: Создает новую запись аналитики для пользователя
      operationId: createUser
      requestBody:
        description: Данные пользователя в формате JSON
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAnalyticsDTO'
      responses:
        '200':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAnalyticsDTO'
        '400':
          description: Некорректные данные пользователя
  /analytics/users/logins:
    post:
      tags:
        - analytics
      summary: Зарегистрировать вход пользователя
      description: Регистрирует факт входа пользователя с заданным ID в систему
      operationId: recordUserLogin
      requestBody:
        description: ID пользователя в формате JSON
        required: true
        content:
          application/json:
            schema:
              type: integer
              format: int64
              example: 12345
      responses:
        '200':
          description: Вход зарегистрирован
          content:
            application/json:
              schema:
                type: string
                format: date-time
        '400':
          description: Некорректные данные
  /analytics/presets/get-preset-by-name/{preset_name}:
    get:
      tags:
        - analytics
      summary: Найти пресет по его названию
      description: Возвращает пресет по указанному названию
      operationId: getPresetByName
      parameters:
        - name: preset_name
          in: path
          description: Название пресета
          required: true
          schema:
            type: string
            example: "Test_Preset"
      responses:
        '200':
          description: Пресет успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresetAnalyticsDTO'
        '404':
          description: Пресет не найден
  /analytics/presets/most-downloaded:
    get:
      tags:
        - analytics
      summary: Получить самый популярный пресет
      description: Возвращает пресет с наибольшим количеством скачиваний
      operationId: getMostDownloadedPreset
      responses:
        '200':
          description: Пресет найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresetAnalyticsDTO'
        '404':
          description: Пресеты не найдены
  /analytics/users/get-user-by-id/{user_id}:
    get:
      tags:
        - analytics
      summary: Найти пользователя по его ID
      description: Возвращает запись о пользователе по указанному ID
      operationId: getUserById
      parameters:
        - name: user_id
          in: path
          description: Идентификатор пользователя
          required: true
          schema:
            type: integer
            format: int64
            example: 12345
      responses:
        '200':
          description: Пользователь успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAnalyticsDTO'
        '404':
          description: Пользователь не найден
  /analytics/users/earliest-login:
    get:
      tags:
        - analytics
      summary: Получить пользователя с самым старым логином
      description: Возвращает пользователя с самым ранним зарегистрированным входом
      operationId: getUserWithEarliestLogin
      responses:
        '200':
          description: Пользователь найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAnalyticsDTO'
        '404':
          description: Пользователи не найдены
  /presets/get-by-id/{id}:
    get:
      tags:
        - presets
      summary: Найти пресет по ID
      description: Возвращает пресет по указанному ID
      operationId: getPresetById
      parameters:
        - name: id
          in: path
          description: ID пресета
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Пресет успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresetDTO'
        '404':
          description: Пресет не найден
  /presets/create:
    post:
      tags:
        - presets
      summary: Создать новый пресет
      description: Создает и возвращает новый пресет
      operationId: createPreset
      requestBody:
        description: Данные пресета в формате JSON
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresetDTO'
      responses:
        '201':
          description: Пресет успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresetDTO'
        '400':
          description: Некорректные данные
  /presets/update-by-id/{id}:
    put:
      tags:
        - presets
      summary: Обновить пресет
      description: Обновляет и возвращает существующий пресет
      operationId: updatePresetById
      parameters:
        - name: id
          in: path
          description: ID пресета
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        description: Обновленные данные пресета
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresetDTO'
      responses:
        '200':
          description: Пресет успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresetDTO'
        '404':
          description: Пресет не найден
        '400':
          description: Некорректные данные
  /presets/delete-by-id/{id}:
    delete:
      tags:
        - presets
      summary: Удалить пресет
      description: Удаляет пресет по указанному ID
      operationId: deletePresetById
      parameters:
        - name: id
          in: path
          description: ID пресета
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '204':
          description: Пресет успешно удален
        '404':
          description: Пресет не найден
...
components:
  schemas:
    PresetAnalyticsDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        presetName:
          type: string
        downloadsNumber:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserAnalyticsDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        loginCount:
          type: integer
          format: int64
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    PresetDTO:
      type: object
      description: Конфигурация пресета настроек окна
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор
          example: 1
          readOnly: true
        preset_name:
          type: string
          description: Название пресета
          example: "Стандартные настройки"
        ventilation_flag:
          type: boolean
          description: Флаг проветривания
          example: true
        ventilation_timer:
          type: integer
          description: Время проветривания (секунды)
          example: 600
          minimum: 0
          maximum: 86400
        darkness_value:
          type: integer
          description: Уровень затемнения (0-100)
          example: 75
          minimum: 0
          maximum: 100
        frame_color:
          type: integer
          description: Цвет подсветки (RGB)
          example: 16777215
      required:
        - preset_name
        - ventilation_flag
        - ventilation_timer
        - darkness_value
        - frame_color
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен для аутентификации
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-refresh
      description: Refresh token в cookie
    api_key:
      type: apiKey
      name: api_key
      in: header